integrated cache with search
